requirements.txt
fastapi
uvicorn
pandas
duckdb
google-genai
python-dotenv
